plant-sentry/server/requirements.txt
Flask==3.1.3
python-dotenv==1.2.2
supabase==2.29.0
